requirements.txt
streamlit==1.28.1
fastapi==0.104.1
pydantic==2.5.0
numpy==1.24.3
scipy==1.11.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
